move apiRequest.js to service directory
